+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2019-05-31
+
+### Fixed
+
+- Border image should be visible now (hardcoded `border-image-slice` value to `1`)
+
+### Removed
+
+- Radial border gradients
+
 ## [2.0.0] - 2019-05-30
 
 ### Added